test: split native runtime descriptor suites
